+++ java/src/main/java/io/cucumber/junitxmlformatter/XmlReportData.java
@@ -16,6 +16,7 @@ import io.cucumber.query.Lineage;
 import io.cucumber.query.NamingStrategy;
 import io.cucumber.query.Query;
 import io.cucumber.query.Repository;
+import org.jspecify.annotations.Nullable;
 
 import java.time.Duration;
 import java.util.AbstractMap.SimpleEntry;
@@ -28,21 +29,25 @@ import java.util.Optional;
 import static io.cucumber.messages.types.TestStepResultStatus.PASSED;
 import static io.cucumber.query.Repository.RepositoryFeature.INCLUDE_GHERKIN_DOCUMENTS;
 import static java.time.format.DateTimeFormatter.ISO_INSTANT;
+import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.SECONDS;
 import static java.util.stream.Collectors.toList;
 
 class XmlReportData {
+    private static final long MILLIS_PER_SECOND = SECONDS.toMillis(1L);
 
     private final Repository repository = Repository.builder()
             .feature(INCLUDE_GHERKIN_DOCUMENTS, true)
             .build();
     private final Query query = new Query(repository);
-    private final NamingStrategy namingStrategy;
+    private final String testSuiteName;
+    private final @Nullable String testClassName;
+    private final NamingStrategy testNamingStrategy;
 
-    private static final long MILLIS_PER_SECOND = SECONDS.toMillis(1L);
-
-    XmlReportData(NamingStrategy namingStrategy) {
-        this.namingStrategy = namingStrategy;
+    XmlReportData(String testSuiteName, @Nullable String testClassName, NamingStrategy testNamingStrategy) {
+        this.testSuiteName = requireNonNull(testSuiteName);
+        this.testClassName = testClassName;
+        this.testNamingStrategy = requireNonNull(testNamingStrategy);
     }
 
     void collect(Envelope envelope) {
@@ -74,20 +79,27 @@ class XmlReportData {
                 .orElseThrow(() -> new IllegalStateException("No pickle for " + testCaseStarted.getId()));
     }
 
-    String getPickleName(TestCaseStarted testCaseStarted) {
+    String getTestName(TestCaseStarted testCaseStarted) {
         Pickle pickle = getPickle(testCaseStarted);
         return query.findLineageBy(pickle)
-                .map(lineage -> namingStrategy.reduce(lineage, pickle))
+                .map(lineage -> testNamingStrategy.reduce(lineage, pickle))
                 .orElseGet(pickle::getName);
     }
 
-    String getFeatureName(TestCaseStarted testCaseStarted) {
+    String getTestClassName(TestCaseStarted testCaseStarted) {
+        if (testClassName != null) {
+            return testClassName;
+        }
         return query.findLineageBy(testCaseStarted)
                 .flatMap(Lineage::feature)
                 .map(Feature::getName)
                 .orElseGet(() -> this.getPickle(testCaseStarted).getUri());
     }
 
+    String getTestSuiteName() {
+        return testSuiteName;
+    }
+    
     List<Entry<String, String>> getStepsAndResult(TestCaseStarted testCaseStarted) {
         return query.findTestStepFinishedAndTestStepBy(testCaseStarted)
                 .stream()
@@ -129,7 +141,7 @@ class XmlReportData {
     }
 
     private static final io.cucumber.messages.types.Duration ZERO_DURATION =
-            new io.cucumber.messages.types.Duration(0L, 0L);
+            new io.cucumber.messages.types.Duration(0L, 0);
     // By definition, but see https://github.com/cucumber/gherkin/issues/11
     private static final TestStepResult SCENARIO_WITH_NO_STEPS = new TestStepResult(ZERO_DURATION, null, PASSED, null);
 
@@ -138,7 +150,7 @@ class XmlReportData {
                 .orElse(SCENARIO_WITH_NO_STEPS);
     }
 
-    public Optional<String> getTestRunStartedAt() {
+    Optional<String> getTestRunStartedAt() {
         return query.findTestRunStarted()
                 .map(TestRunStarted::getTimestamp)
                 .map(Convertor::toInstant)