+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed 
 - Update dependency io.cucumber:messages up to v30
 
+## [0.11.0] - 2025-11-14
+### Added
+- Support configurable testsuite name, testcase classname, and testcase name. ([#105](https://github.com/cucumber/junit-xml-formatter/pull/105))
+
+## [0.10.0] - 2025-10-27
+### Changed
+- Update dependency io.cucumber:messages up to v30
+
 ## [0.9.0] - 2025-09-11
 ### Changed
 - Update dependency cucumber/query  to 14.0.1
@@ -67,7 +75,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Java implementation ([#3](https://github.com/cucumber/junit-xml-formatter/pull/3) M.P. Korstanje)
 
-[Unreleased]: https://github.com/cucumber/junit-xml-formatter/compare/v0.9.0...HEAD
+[Unreleased]: https://github.com/cucumber/junit-xml-formatter/compare/v0.11.0...HEAD
+[0.11.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.10.0...v0.11.0
+[0.10.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/cucumber/junit-xml-formatter/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/cucumber/junit-xml-formatter/compare/v0.7.1...v0.8.0